core/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
databases[aiosqlite]>=0.8.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
asgi-correlation-id>=4.3.0
python-json-logger>=2.0.7
rich>=13.7.0
